pyproject.toml
[tool.poetry]
name = "snowbird"
version = "0.0.1"
description = "Snowbird configures Snowflake resources for dataproducts. Builds on Permifrost." 
authors = ["pbencze <paul@idelab.com>"]

[tool.poetry.dependencies]
python = "^3.8"
permifrost = "^0.14.0"
pydantic = "^1.10.2"
PyYAML = "^6.0"
click = "^8.1.3"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
mock = "^4.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
snowbird = 'snowbird.cli:cli'
